fix: Introduce newHost input
Because 200+ repos use `sonar.ts.sv` as the value of the `host` input. That URL is
no longer good and working, so insted of raising 200+ PRs to update it, we introduce
a new input called `newHost` with a working URL as the default value, and the action's
code will only read `newHost` input and ignore the `host` input.